Separate components from App
